chore: refactor config resolution module and update imports
